refactor: Added useDocumentationSearch hook, moved logic to it and used it under DocumentationSearch component - Added search dismiss icon - Moved DocumentationSearch to its own directory
