+++ app/routes/_landing.($lang).docs/components/DocumentationSearch/index.ts
@@ -1,0 +1,3 @@
+import DocumentationSearch from "./DocumentationSearch"
+export * from "./DocumentationSearch"
+export default DocumentationSearch
