+++ app/routes/_landing.($lang).docs/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./useDocumentationSearch"
